+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `openeo.udf.run_code.extract_udf_dependencies()` to extract UDF dependency declarations from UDF code
+  (related to [Open-EO/openeo-geopyspark-driver#237](https://github.com/Open-EO/openeo-geopyspark-driver/issues/237))
+
 ### Changed
 
 ### Removed